+++ src/Microsoft.VisualStudio.Editor.Razor/AttributeCompletionResult.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNetCore.Razor.Language;
 
-namespace Microsoft.CodeAnalysis.Razor
+namespace Microsoft.VisualStudio.Editor.Razor
 {
     public abstract class AttributeCompletionResult
     {
