+++ src/Microsoft.VisualStudio.Editor.Razor/TagHelperCompletionService.cs
@@ -1,11 +1,9 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis.Host;
-
-namespace Microsoft.CodeAnalysis.Razor
+namespace Microsoft.VisualStudio.Editor.Razor
 {
-    public abstract class TagHelperCompletionService : ILanguageService
+    public abstract class TagHelperCompletionService
     {
         public abstract AttributeCompletionResult GetAttributeCompletions(AttributeCompletionContext completionContext);
